backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
passlib==1.7.4
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.9.0
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.12.2
uvicorn==0.33.0
